Move ResizeTooltip to ResizableBox. Enable with __experimental. Remove export.
